[GSCOLLECT-780] Create interfaces for primitive sets, bags, stacks, maps.
git-svn-id: svn+ssh://gscollections.svn.services.gs.com/svnroot/gscollections-svn/trunk@133 d5c9223b-1aff-41ac-aadd-f810b4a99ac4

+++ gs-collections-code-generator/src/main/java/com/gs/collections/codegenerator/tools/FileUtils.java
@@ -18,8 +18,12 @@ package com.gs.collections.codegenerator.tools;
 
 import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.nio.MappedByteBuffer;
+import java.nio.channels.FileChannel;
+import java.nio.charset.Charset;
 import java.util.List;
 
 public final class FileUtils
@@ -31,41 +35,44 @@ public final class FileUtils
 
     public static void writeToFile(String data, File outputFile)
     {
-        FileWriter fileWriter = null;
-        BufferedWriter bufferedWriter = null;
-        try
+        if (!outputFile.exists() || (outputFile.exists() && !readFile(outputFile.getAbsolutePath()).equals(data)))
         {
-            fileWriter = new FileWriter(outputFile);
-            bufferedWriter = new BufferedWriter(fileWriter);
-            bufferedWriter.write(data);
-            bufferedWriter.flush();
-        }
-        catch (IOException e)
-        {
-            throw new RuntimeException("Could not write generated sources to file: " + e);
-        }
-        finally
-        {
-            if (fileWriter != null)
+            FileWriter fileWriter = null;
+            BufferedWriter bufferedWriter = null;
+            try
             {
-                try
-                {
-                    fileWriter.close();
-                }
-                catch (IOException e)
-                {
-                    throw new RuntimeException("Could not close filewriter: " + e);
-                }
+                fileWriter = new FileWriter(outputFile);
+                bufferedWriter = new BufferedWriter(fileWriter);
+                bufferedWriter.write(data);
+                bufferedWriter.flush();
             }
-            if (bufferedWriter != null)
+            catch (IOException e)
             {
-                try
+                throw new RuntimeException("Could not write generated sources to file: " + e);
+            }
+            finally
+            {
+                if (fileWriter != null)
                 {
-                    bufferedWriter.close();
+                    try
+                    {
+                        fileWriter.close();
+                    }
+                    catch (IOException e)
+                    {
+                        throw new RuntimeException("Could not close filewriter: " + e);
+                    }
                 }
-                catch (IOException e)
+                if (bufferedWriter != null)
                 {
-                    throw new RuntimeException("Could not close bufferedwriter: " + e);
+                    try
+                    {
+                        bufferedWriter.close();
+                    }
+                    catch (IOException e)
+                    {
+                        throw new RuntimeException("Could not close bufferedwriter: " + e);
+                    }
                 }
             }
         }
@@ -100,5 +107,38 @@ public final class FileUtils
                 throw new RuntimeException("Could not create directory " + path);
             }
         }
+    }
+
+    private static String readFile(String path)
+    {
+        FileInputStream stream = null;
+        FileChannel fc;
+        MappedByteBuffer bb = null;
+        try
+        {
+            stream = new FileInputStream(new File(path));
+            fc = stream.getChannel();
+            bb = fc.map(FileChannel.MapMode.READ_ONLY, 0, fc.size());
+        }
+        catch (IOException e)
+        {
+            throw new RuntimeException(e);
+        }
+        finally
+        {
+            try
+            {
+                if (stream != null)
+                {
+                    stream.close();
+                }
+            }
+            catch (IOException e)
+            {
+                throw new RuntimeException(e);
+            }
+        }
+        /* Instead of using default, pass in a decoder. */
+        return Charset.defaultCharset().decode(bb).toString();
     }
 }
